+++ example/lib/advanced/base_consumer_mock.dart
@@ -22,47 +22,18 @@ abstract class BaseConsumerMock<T extends AbstractModel<ID>, ID>
   Future<ConsumerPermission> checkPermission(
     BuildContext context,
     List<String>? paths,
-  ) {
-    // if (paths?.join('/').contains('example_map_function_route') ?? false) {
-    //   return Future<ConsumerPermission>.value(
-    //     const ConsumerPermission(
-    //       name: 'Example Map Func. Route',
-    //       iconName: 'cube',
-    //       view: true,
-    //       insert: true,
-    //       update: true,
-    //       delete: true,
-    //       menu: true,
-    //     ),
-    //   );
-    // }
-
-    // if (paths?.join('/').contains('example_model_function_route') ?? false) {
-    //   return Future<ConsumerPermission>.value(
-    //     const ConsumerPermission(
-    //       name: 'Example Model Func. Route',
-    //       iconName: 'mugHot',
-    //       view: true,
-    //       insert: true,
-    //       update: true,
-    //       delete: true,
-    //       menu: true,
-    //     ),
-    //   );
-    // }
-
-    return Future<ConsumerPermission>.value(
-      const ConsumerPermission(
-        name: 'mock',
-        iconName: 'question',
-        view: true,
-        insert: true,
-        update: true,
-        delete: true,
-        menu: true,
-      ),
-    );
-  }
+  ) =>
+      Future<ConsumerPermission>.value(
+        const ConsumerPermission(
+          name: 'mock',
+          iconName: 'question',
+          view: true,
+          insert: true,
+          update: true,
+          delete: true,
+          menu: true,
+        ),
+      );
 
   ///
   ///
@@ -127,10 +98,10 @@ abstract class BaseConsumerMock<T extends AbstractModel<ID>, ID>
   ///
   ///
   @override
-  Future<bool> saveOrUpdate(
+  Future<ID?> saveOrUpdate(
     BuildContext context,
     T model, {
     Map<String, String> extraParams = const <String, String>{},
   }) async =>
-      true;
+      model.id;
 }